requirements.txt
pyTelegramBotAPI>=4.0.0
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0